requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
anyio==4.11.0
asyncio==4.0.0
attrs==25.4.0
bcrypt==5.0.0
beautifulsoup4==4.14.2
Brotli==1.1.0
bs4==0.0.2
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
colorama==0.4.6
cryptography==46.0.3
cssselect2==0.8.0
defusedxml==0.7.1
dnspython==2.8.0
filelock==3.20.0
fonttools==4.60.1
fpdf2==2.8.5
frozenlist==1.8.0
fsspec==2025.10.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
invoke==2.2.1
Jinja2==3.1.6
ldap3==2.9.1
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.0
networkx==3.5
numpy==1.24.3
outcome==1.3.0.post0
packaging==25.0
pandas==2.0.3
paramiko==4.0.0
pdfkit==1.0.0
pillow==12.0.0
pluggy==1.6.0
propcache==0.4.1
psycopg2==2.9.11
pyasn1==0.6.1
pycparser==2.23
pydyf==0.11.0
Pygments==2.19.2
pymongo==4.15.3
PyMySQL==1.1.2
PyNaCl==1.6.0
pyphen==0.17.2
PySocks==1.7.1
pytest==8.4.2
python-dateutil==2.9.0.post0
python-nmap==0.7.1
pytz==2025.2
PyYAML==6.0.1
redis==7.0.1
reportlab==4.0.4
requests==2.31.0
rich==13.5.0
scapy==2.5.0
schedule==1.2.0
selenium==4.15.0
six==1.17.0
sniffio==1.3.1
socksio==1.0.0
sortedcontainers==2.4.0
soupsieve==2.8
sqlmap==1.9.10
stem==1.8.2
subprocess32==3.5.4
sympy==1.14.0
tinycss2==1.4.0
tinyhtml5==2.0.0
torch==2.1.0
torrequest==0.1.0
trio==0.32.0
trio-websocket==0.12.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
weasyprint==66.0
webencodings==0.5.1
wsproto==1.2.0
yarl==1.22.0
zopfli==0.2.3.post1